non working useMemo when calling useContext
