+++ capstone/src/Contexts/AdventurerContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const AdventurerContext = createContext();
+
+export default AdventurerContext;
