+++ capstone/src/components/User/Adventurer/Adventurer.css
@@ -4,6 +4,6 @@
   grid-template-columns: 2fr 1fr;
 }
 
-.adventure .profile {
+.adventure .experienceInfo {
   text-align: center;
 }
