+++ capstone/src/components/User/Adventurer/Adventurer.jsx
@@ -1,27 +1,74 @@
 import './Adventurer.css';
 import { useNavigate } from 'react-router-dom';
-import { useEffect } from 'react';
+import { useEffect, useState, useMemo } from 'react';
+import { axios } from 'axios';
 import FindAdventure from './FindAdventure/FindAdventure';
 import Header from '../Header/Header';
 
+// CONTEXTS
+import AdventurerContext from '../../../Contexts/AdventurerContext';
+
+const baseURL = process.env.REACT_APP_BASE_URL;
+
+function getNearbyRestaurants(location) {
+  return axios.post(`${baseURL}/adventure/restaurants`, location);
+}
+
+// things map needs:
+// restaurants
+// isDataFetched
+// currentPosition
 export default function Adventurer({ setIsLoggedIn, isLoggedIn }) {
+  // STATES
+  const [currentPosition, setCurrentPosition] = useState({ lat: 0, lng: 0 });
+  const [isDataFetched, setIsDataFetched] = useState(false);
+  const [restaurants, setRestaurants] = useState([]);
+  // MEMO VALUES
+  const memoPosition = useMemo(() => ({ value: currentPosition }), [currentPosition]);
+  const memoDataFetched = useMemo(() => ({ value: isDataFetched }), [isDataFetched]);
+  const memoRestaurants = useMemo(() => ({ value: restaurants }), [restaurants]);
+
+  // Check if user is logged in
   const navigate = useNavigate();
   useEffect(() => {
     if (!isLoggedIn) {
       navigate('/');
     }
   }, [isLoggedIn]);
+  // Find current position
+  useEffect(() => {
+    navigator.geolocation.getCurrentPosition((position) => {
+      setCurrentPosition({ lat: position.coords.latitude, lng: position.coords.longitude });
+    });
+  }, [isLoggedIn]);
+  // Fetch closest restaurants
+  useEffect(() => {
+    if (currentPosition.lat === 0 && currentPosition.lng === 0) {
+      setRestaurants([]);
+    } else {
+      getNearbyRestaurants(currentPosition)
+        .then((res) => setRestaurants(res.data))
+        .catch(console.error);
+      setIsDataFetched(true);
+    }
+  }, [currentPosition, setCurrentPosition, setRestaurants]);
   return (
     <div className="user">
       <Header onLogOutClick={setIsLoggedIn} />
-      <div className="adventure">
-        <FindAdventure
-          isLoggedIn={isLoggedIn}
-        />
-        <div className="profile">
-          <h2>History</h2>
+      <AdventurerContext.Provider
+        value={{
+          currentPosition: memoPosition,
+          isDataFetched: memoDataFetched,
+          restaurants: memoRestaurants,
+        }}
+      >
+        <div className="adventure">
+          <FindAdventure />
+          <div className="experienceInfo">
+            <h2>Experience Information</h2>
+          </div>
         </div>
-      </div>
+      </AdventurerContext.Provider>
     </div>
   );
 }
